# Use the most space-efficient encoding for dense stores