--- a/src/main/java/com/datadoghq/sketch/ddsketch/encoding/VarEncodingHelper.java
+++ b/src/main/java/com/datadoghq/sketch/ddsketch/encoding/VarEncodingHelper.java
@@ -12,6 +12,37 @@ public final class VarEncodingHelper {
   private static final int MAX_VAR_LEN_64 = 9;
   private static final int VAR_DOUBLE_ROTATE_DISTANCE = 6;
 
+  private static final byte[] UNSIGNED_VAR_LONG_LENGTHS = new byte[65];
+  private static final byte[] VAR_DOUBLE_LENGTHS = new byte[65];
+
+  static {
+    final GrowingByteArrayOutput output =
+        GrowingByteArrayOutput.withInitialCapacity(MAX_VAR_LEN_64);
+
+    for (int numLeadingZeros = 0; numLeadingZeros <= 64; numLeadingZeros++) {
+      output.clear();
+      try {
+        encodeUnsignedVarLong(output, numLeadingZeros == 64 ? 0L : -1L >>> numLeadingZeros);
+      } catch (IOException e) {
+        // Propagate the exception.
+        throw new RuntimeException(e);
+      }
+      UNSIGNED_VAR_LONG_LENGTHS[numLeadingZeros] = (byte) output.trimmedCopy().length;
+    }
+
+    for (int numTrailingZeros = 0; numTrailingZeros <= 64; numTrailingZeros++) {
+      output.clear();
+      try {
+        encodeVarDouble(
+            output, varBitsToDouble(numTrailingZeros == 64 ? 0L : -1L << numTrailingZeros));
+      } catch (IOException e) {
+        // Propagate the exception.
+        throw new RuntimeException(e);
+      }
+      VAR_DOUBLE_LENGTHS[numTrailingZeros] = (byte) output.trimmedCopy().length;
+    }
+  }
+
   private VarEncodingHelper() {}
 
   /**
@@ -54,6 +85,18 @@ public final class VarEncodingHelper {
   }
 
   /**
+   * {@code unsignedVarLongEncodedLength} returns the number of bytes that {@link
+   * #encodeUnsignedVarLong(Output, long)} encodes a {@code long} value into.
+   *
+   * @param value the value to encode
+   * @return the number of bytes that {@link #encodeUnsignedVarLong(Output, long)} encodes a {@code
+   *     long} value into
+   */
+  public static byte unsignedVarLongEncodedLength(long value) {
+    return UNSIGNED_VAR_LONG_LENGTHS[Long.numberOfLeadingZeros(value)];
+  }
+
+  /**
    * {@code encodeSignedVarLong} serializes {@code long} values using zig-zag encoding, which
    * ensures small-scale integers are turned into integers that have leading zeros, whether they are
    * positive or negative, hence allows for space-efficient encoding of those values.
@@ -63,7 +106,7 @@ public final class VarEncodingHelper {
    * @throws IOException if an {@link IOException} is thrown while writing to {@code output}
    */
   public static void encodeSignedVarLong(final Output output, final long value) throws IOException {
-    encodeUnsignedVarLong(output, (value >> (64 - 1) ^ (value << 1)));
+    encodeUnsignedVarLong(output, zigZagEncode(value));
   }
 
   /**
@@ -75,7 +118,26 @@ public final class VarEncodingHelper {
    * @throws IOException if an {@link IOException} is thrown while reading from {@code input}
    */
   public static long decodeSignedVarLong(final Input input) throws IOException {
-    final long value = decodeUnsignedVarLong(input);
+    return zigZagDecode(decodeUnsignedVarLong(input));
+  }
+
+  /**
+   * {@code signedVarLongEncodedLength} returns the number of bytes that {@link
+   * #encodeSignedVarLong(Output, long)} encodes a {@code long} value into.
+   *
+   * @param value the value to encode
+   * @return the number of bytes that {@link #encodeSignedVarLong(Output, long)} encodes a {@code
+   *     long} value into
+   */
+  public static byte signedVarLongEncodedLength(long value) {
+    return UNSIGNED_VAR_LONG_LENGTHS[Long.numberOfLeadingZeros(zigZagEncode(value))];
+  }
+
+  private static long zigZagEncode(long value) {
+    return value >> (64 - 1) ^ (value << 1);
+  }
+
+  private static long zigZagDecode(long value) {
     return (value >>> 1) ^ -(value & 1);
   }
 
@@ -102,10 +164,7 @@ public final class VarEncodingHelper {
    * @throws IOException if an {@link IOException} is thrown while writing to {@code output}
    */
   public static void encodeVarDouble(final Output output, final double value) throws IOException {
-    long bits =
-        Long.rotateLeft(
-            Double.doubleToRawLongBits(value + 1) - Double.doubleToRawLongBits(1),
-            VAR_DOUBLE_ROTATE_DISTANCE);
+    long bits = doubleToVarBits(value);
     for (int i = 0; i < MAX_VAR_LEN_64 - 1; i++) {
       final byte next = (byte) (bits >>> (8 * 8 - 7));
       bits <<= 7;
@@ -140,6 +199,28 @@ public final class VarEncodingHelper {
       }
       bits |= ((long) next & 0x7FL) << shift;
     }
+    return varBitsToDouble(bits);
+  }
+
+  /**
+   * {@code varDoubleEncodedLength} returns the number of bytes that {@link #encodeVarDouble(Output,
+   * double)} encodes a {@code double} value into.
+   *
+   * @param value the value to encode
+   * @return the number of bytes that {@link #encodeVarDouble(Output, double)} encodes a {@code
+   *     double} value into
+   */
+  public static byte varDoubleEncodedLength(double value) {
+    return VAR_DOUBLE_LENGTHS[Long.numberOfTrailingZeros(doubleToVarBits(value))];
+  }
+
+  private static long doubleToVarBits(double value) {
+    return Long.rotateLeft(
+        Double.doubleToRawLongBits(value + 1) - Double.doubleToRawLongBits(1),
+        VAR_DOUBLE_ROTATE_DISTANCE);
+  }
+
+  private static double varBitsToDouble(long bits) {
     return Double.longBitsToDouble(
             Long.rotateRight(bits, VAR_DOUBLE_ROTATE_DISTANCE) + Double.doubleToRawLongBits(1))
         - 1;