When serializing a dense store with the custom encoding, figure out if the data would be more space-efficiently encoded densely or sparsely, and use the better option.

This is similar to what was done in https://github.com/DataDog/sketches-go/pull/53.

# Benchmark

As in the Go implementation, there is a computational cost. See the benchmarks in https://github.com/DataDog/sketches-go/pull/53 for the gain in space efficiency.

## Before

```
Serialize.encode          100000                    POISSON                0.01            FAST  NANOSECONDS  avgt    3  2.617 ± 0.748  us/op
Serialize.encode          100000  COMPOSITE_POISSON_EXTREME                0.01            FAST  NANOSECONDS  avgt    3  4.877 ± 5.788  us/op
Serialize.encode          100000            TRIMODAL_NORMAL                0.01            FAST  NANOSECONDS  avgt    3  3.995 ± 1.709  us/op
Serialize.encodeReusing   100000                    POISSON                0.01            FAST  NANOSECONDS  avgt    3  1.293 ± 0.212  us/op
Serialize.encodeReusing   100000  COMPOSITE_POISSON_EXTREME                0.01            FAST  NANOSECONDS  avgt    3  3.661 ± 0.521  us/op
Serialize.encodeReusing   100000            TRIMODAL_NORMAL                0.01            FAST  NANOSECONDS  avgt    3  2.236 ± 0.117  us/op
```

## After

```
Serialize.encode          100000                    POISSON                0.01            FAST  NANOSECONDS  avgt    3  2.879 ± 0.916  us/op
Serialize.encode          100000  COMPOSITE_POISSON_EXTREME                0.01            FAST  NANOSECONDS  avgt    3  6.311 ± 0.581  us/op
Serialize.encode          100000            TRIMODAL_NORMAL                0.01            FAST  NANOSECONDS  avgt    3  4.987 ± 1.020  us/op
Serialize.encodeReusing   100000                    POISSON                0.01            FAST  NANOSECONDS  avgt    3  2.277 ± 0.275  us/op
Serialize.encodeReusing   100000  COMPOSITE_POISSON_EXTREME                0.01            FAST  NANOSECONDS  avgt    3  5.304 ± 1.173  us/op
Serialize.encodeReusing   100000            TRIMODAL_NORMAL                0.01            FAST  NANOSECONDS  avgt    3  4.226 ± 0.443  us/op
```